Add the ability to manually override fill precedence, plus tests
We had this before, I removed it for some reason.

Added it back to solve the problem with child materials all getting the
same pad size if they're all shrinky, resulting in weird behavior if you
want, for example, an hbox with a fixed size sidebar, and literally
anything non-greedy in the second row.

I think I thought just using free-layout or some sort of special greedy
container was more elegant... I disagree at this juncture.