Add sanity check for graceful failure
Was causing a crash when an unpositioned element had an invalid canvas.
The reason it never happened with invalid geometry is because std::map
implicitly created a pair for it because it used the subscript operator,
whereas invalidateCanvas used the at() function. I made the check more
explicit in the former case.