Make general performance with tons of items orders of magnitude faster.
With this one weird trick.

GAHHH. I don't regret the attribute rewrite, it was necessary for
speeding up things in general, and it's SOO much nicer to reason about now,
but I think this was actually the problem that I was looking for when I
started it, where I have 300 model items spun up and it just ground to a
screaching hault. Now it's not exactly snappy, but it looks like the
majority of time is now spent in ::present(), which I can totally believe...

Time for progressive redrawing? :3c
I thought before maybe that was overkill on modern computers but I guess
software rendering just sucks period. Fullscreen'll be something like 4 times slower in 4K, and this window I'm testing is only 640x400......

Edit: Yeah, we need progressive updates. This is just too slow...