Port rowlayout -> boxlayout **
** Concerned about whether the line shaping function was the latest
version I did. There's about a week of work I did on IVD last summer on
my other computer I don't have access to yet, and I don't remember what
it was but there are decent odds it was this function. Oh well, it's
easy enough to port the other version back in.